+++ src/Hacks/Hitboxes.cpp
@@ -1,0 +1,14 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/GJBaseGameLayer.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+class $modify(GJBaseGameLayer) {
+    virtual void update(float p0)
+    {
+        GJBaseGameLayer::update(p0);
+
+        
+    }
+};